requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
pydantic
python-dotenv
azure-servicebus==7.12.1
pytest>=7.0
httpx
